Founded Items structure
Now we have Items!
Item is scriptableObject so you can add yours by adding an item in your prefab\items folder, just like any other asset.
if you want a couple of items together, you're going have to need ItemStack class which is already implemented.
You can add a stone on 'e' btn as a test object.

ToDo:
When we add a couple of items we take just as much as we can. But we need to let external storage know how much did we get, or how much remains in storage. And also we need to implement storages, yisss ^w^